Rename opendg->tgm
