+++ .pre-commit-config.yaml
@@ -75,7 +75,7 @@ repos:
     hooks:
       - id: mypy
         name: mypy
-        entry: uv run mypy opendg
+        entry: uv run mypy tgm
         language: system
         always_run: true
         pass_filenames: false
