+++ examples/nodeproppred/gcn.py
@@ -9,10 +9,10 @@ from tgb.nodeproppred.evaluate import Evaluator
 from torch_geometric.nn import GCNConv
 from tqdm import tqdm
 
-from opendg.graph import DGBatch, DGraph
-from opendg.loader import DGDataLoader
-from opendg.timedelta import TimeDeltaDG
-from opendg.util.seed import seed_everything
+from tgm.graph import DGBatch, DGraph
+from tgm.loader import DGDataLoader
+from tgm.timedelta import TimeDeltaDG
+from tgm.util.seed import seed_everything
 
 parser = argparse.ArgumentParser(
     description='GCN Example',
