--- opendg/__init__.py
@@ -1,7 +1,0 @@
-"""OpenDG library for fast ML on temporal graphs."""
-
-__version__ = '0.1.0'
-
-__all__ = [
-    '__version__',
-]
