+++ tgm/__init__.py
@@ -1,0 +1,7 @@
+"""TGM: Efficient and Modular Training on Dynamic Graphs."""
+
+__version__ = '0.0.1-alpha'
+
+__all__ = [
+    '__version__',
+]
