+++ tgm/_storage/backends/__init__.py
@@ -1,4 +1,4 @@
-from opendg._storage.backends.array_backend import DGStorageArrayBackend
+from tgm._storage.backends.array_backend import DGStorageArrayBackend
 
 DGStorageBackends = {
     'ArrayBackend': DGStorageArrayBackend,
