+++ tgm/hooks.py
@@ -5,8 +5,8 @@ from typing import Any, Deque, Dict, List, Protocol, Set, runtime_checkable
 
 import torch
 
-from opendg._storage import DGSliceTracker
-from opendg.graph import DGBatch, DGraph
+from tgm._storage import DGSliceTracker
+from tgm.graph import DGBatch, DGraph
 
 
 @runtime_checkable
