+++ tgm/nn/memory/edgebank.py
