+++ tgm/nn/time_encoding.py
