+++ tgm/timedelta.py
