+++ tgm/util/perf.py
