+++ tgm/util/seed.py
